backend/requirements.txt
# [JIRA編]React Hooks/TypeScript + Django REST APIで作るオリジナルJIRA を参考
Django==3.1
djangorestframework==3.11.1
django-cors-headers==3.4.0
djangorestframework-simplejwt==4.6.0
djoser==2.0.3
pillow==7.2.0
psycopg2-binary==2.8.6


# 現場で使えるDjangoの教科書より
django-environ